requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
python-dotenv==0.10.1
requests==2.21.0